+++ packages/nextjs/app/_components/ChallengeExpandedCard.tsx
@@ -1,10 +1,10 @@
 import Image from "next/image";
 import Link from "next/link";
+import { getChallengeDependenciesInfo } from "../../utils/dependent-challenges";
 import CrossedSwordsIcon from "../_assets/icons/CrossedSwordsIcon";
 import PadLockIcon from "../_assets/icons/PadLockIcon";
 import QuestionIcon from "../_assets/icons/QuestionIcon";
 import { ChallengeData, challengesData } from "../_data/_hardcoded";
-import { getChallengeDependenciesInfo } from "./utils";
 import { ReviewAction } from "~~/services/database/config/types";
 import { getChallengeById } from "~~/services/database/repositories/challenges";
 import { UserChallenges } from "~~/services/database/repositories/userChallenges";
