+++ packages/nextjs/utils/dependent-challenges.ts
